requirements.txt
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
requests==2.31.0
supabase==1.0.0
python-dotenv==1.0.0
flask-wtf==1.1.1
wtforms==3.0.1
validators==0.20.0

